Move tensor4 into cubatic, make operators members
